Cleanup - reorganize Resources files
