+++ core/lib/resources/message_list/index.js
